Send command before device is connected
By default the app tries to connect with the device and aborts if it can't
connect. But you might want to start the app, have it wait until the device is
connected and then send the command.

```
> inputmodule-control b1-display --pattern black
Failed to find serial devivce. Please manually specify with --serial-dev

> inputmodule-control --wait-for-device b1-display --pattern black

> inputmodule-control --wait-for-device b1-display --pattern black
Device already present. No need to wait. Not executing command.
```

Signed-off-by: Daniel Schaefer <dhs@frame.work>