ledmatrix: Limit max brightness, and do custom brightness control
The new hardware is extremely bright, so we need to limit the
brightness.
We limit it so that the max power consumption is slightly under 500mA.
This limit is done by using the scaling factor in the LED controller.

This also has the advantage that we have more brightness levels
available for each LED because we can use the full range (previously
much of the range would be too bright to use, even with the first
hardware revision).

But since we can't use the scaling factor for brightness control
anymore, we have to do the brightness calculation ourselves.

Signed-off-by: Daniel Schaefer <dhs@frame.work>