windows: Statically link vcruntime DLL
Otherwise it won't run on some freshly installed machines.
Windows complains that `vcruntime140.dll` is missing.
Users would have to install the MSVC C++ Redistributable package.

This is the same fix as in FrameworkComputer/qmk_hid.
The measurements of that binary are:
On a local build of `cargo clean; cargo build --release`:
Size of .exe before: 857 KB
Size of .exe after: 878 KB

Note: Only affects release builds! I think that's because the debug version of this DLL isn't redistributable.

Signed-off-by: Daniel Schaefer <dhs@frame.work>